Add files via upload
This project is to design a suitable network system for universities in developing countries. The aim was to design a network with high security and low cost. This project will help to enhanced education of developing countries. The advantages of networking can be seen clearly in terms of efficiency, security, manageability and cost as it allows collaboration between users in a wide area. To improve college campus network design, the technology used was creating LAN, WLAN and using cheap device to reduce cost of the network. But the network can also become better using routing protocols and other protocol. So, we are going to use such protocols using less number of devices and will also maintain the cost of the network less. To design such network, we are going to use software Cisco-Packet Tracer.


Networking is refereed as connecting computers electronically for the purpose of sharing information. Resources such as a file, applications, printers & software are some common information shared in a networking. The advantages of networking can be seen clearly in terms of security, efficiency, manageability & and cost effectiveness as in allows collaboration between users in a wide range. The Switches and Router this device that play an important role in data transfer from one place to another using different technology such as a radio waves & wire.LAN network is made up of two or more computers connected together in a short distance usually at home, offices buildings or school. WAN is a network that covers wider area than LAN and usually covers cities, countries and the whole world. Several major LAN can be connected together to form a WAN. As a several devices are connected to network, it is important to ensure data collision does not happen when this device attempt to use data channel simultaneously. A set of rules called carrier sense multiple access/collision detection are used to detect and prevent collision in networks.

In this growing network area, it has become necessary to protect our network from unauthorized users and prevent it from hacking, so it is necessary to maintain security in our network by using various security option like port security, encryption using most secure routing protocol. To implement this, we need best devices which can support these protocols more efficiently. So, our discussion we decide to use layer 3 switches which work as switch as well as router, and using this router it is possible to implement EIGRP routing protocol. By using EIGRP it is possible for load balancing on parallel links between sites and also manages load balancing.

This project has proven that a standard network system can be designed with less cost. Although we used the cheapest devices in designing the network, the security of this network turned out to be verystrong. This is because the firewall and backup devices used in this network are of good quality. All networks need many servers for doing their work. For this research, we did not use all servers because of cost, but we used some important servers such as DHCP. These servers help the network to perform their functions in a smooth way. It can be seen in this research that various costs were minimized in order to maximize the quality of the designed network. Although there may have been some challenges in this project due to some financial constraints, at the end our aim was achieved by designing a network for developing universities with minimal cost. For example, we made use of some devices for the network security, but the most interesting part is that, at the end of the day, all challenge and constraints were overcome.